# api/swagger.yaml
openapi: 3.0.1
info:
  title: Gortas Authentication Service
  description: ''
  termsOfService: 'TODO'
  contact:
    email: 'maxim.thomas@gmail.com'
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
servers:
  - url: https://gortas:8443/v1/
tags:
  - name: authentication
    description: Authentication service
    externalDocs:
      description: Find out more
      url: http://swagger.io
paths:
  /auth/{realm}/{flow}:
    get:
      tags:
        - authentication
      summary: start authentication
      operationId: startAuth
      parameters:
        - name: realm
          in: path
          description: Realm to authenticate
          required: true
          schema:
            type: string
        - name: flow
          in: path
          description: Authentication service
          required: true
          schema:
            type: string
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CredentialsRequest'
    post:
      tags:
        - authentication
      summary: submit credentials
      operationId: submit authentication data
      parameters:
        - name: realm
          in: path
          description: Realm to authenticate
          required: true
          schema:
            type: string
        - name: flow
          in: path
          description: Service ti authenticate to authenticate
          required: true
          schema:
            type: string
      requestBody:
        description: Credentials data
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CredentialsResponse'
        required: true
      responses:
        405:
          description: Invalid input
          content: {}

components:
  schemas:
    Credential:
      type: object
      properties:
        type:
          type: string
        value:
          type: string
        validation:
          type: string
          description: 'Regular expression to validate field'
        required:
          type: boolean
        properties:
          additionalProperties:
            type: object        
        error: 
          type: string
          description: 'Error message if input is incorrect'
      required:
        - type
        - value
    CredentialsRequest:
      type: object
      properties:
        flowId:
          type: string
        module:
          type: string
        credentials:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/Credential'
    CredentialsResponse:
      allOf:
        - $ref: '#/components/schemas/CredentialsRequest'